virtual_network: Add connectivity test for vdpa interface
1. Add connectivity test for vdpa interface into virtual network
2. Update to ping gateway due to network changes
3. Remove redundant files in virtual_interface
