--- libvirt/tests/cfg/virtual_interface/connectivity_check.cfg
@@ -1,39 +1,0 @@
-- iface.connectivity_check:
-    type = connectivity_check
-    start_vm = no
-
-    variants dev_type:
-        - vdpa:
-            only x86_64
-            func_supported_since_libvirt_ver = (7, 3, 0)
-            func_supported_since_qemu_kvm_ver = (6, 0, 0)
-            iface_dict = {"source": {'dev':'/dev/vhost-vdpa-0'}}
-            save_error = "yes"
-            vm_iface_driver = "virtio_net"
-            variants test_target:
-                - simulator:
-                - mellanox:
-            variants:
-                - @default:
-                - driver_queues:
-                    only mellanox
-                    required_kernel = [5.14.0,)
-                    func_supported_since_libvirt_ver = (8, 0, 0)
-                    driver_queues = 8
-                    vdpa_mgmt_tool_extra = "max_vqp ${driver_queues}"
-                    iface_dict = {'model': 'virtio', 'source': {'dev': '/dev/vhost-vdpa-0'}, 'type_name': 'vdpa', 'driver': {'driver_attr': {'queues': '${driver_queues}'}}, 'mac_address': mac_addr}
-                    variants:
-                        - vcpu_gt_queues:
-                            vm_attrs = {'vcpu': 4}
-                        - vcpu_lt_queues:
-                            vm_attrs = {'vcpu': 16}
-                        - boot:
-                            iface_dict = {'model': 'virtio', 'source': {'dev': '/dev/vhost-vdpa-0'}, 'type_name': 'vdpa', 'driver': {'driver_attr': {'queues': '${driver_queues}'}}, 'boot': '2', 'mac_address': mac_addr}
-                            disk_boot = 1
-                        - page_per_vq:
-                            enable_guest_iommu = "yes"
-                            expr_multiplier = '00001000'
-                            iface_dict = {'model': 'virtio', 'source': {'dev': '/dev/vhost-vdpa-0'}, 'type_name': 'vdpa', 'driver': {'driver_attr': {'queues': '${driver_queues}', 'page_per_vq': 'on'}}, 'mac_address': mac_addr}
-                        - multi_ifaces:
-                            func_supported_since_libvirt_ver = (8, 7, 0)
-                            iface_dict2 = {'model': 'virtio', 'source': {'dev': '/dev/vhost-vdpa-1'}, 'type_name': 'vdpa', 'driver': {'driver_attr': {'queues': '${driver_queues}'}}}
